net: posix_server_socket_impl: implement proxy protocol in server sockets
Implement the proxy protocol v2 as defined by haproxy. This is a header
prepended to the input stream containing the "real" local and remote
addresses, hidden by the reverse proxy or load balancer.

This cannot be implemented by the application, because we want
port load balancing (load_balancing_algorithm::port) to continue
to work and use the client's source port. Connection migration can
only be done by Seastar at the moment.

We can't use input_stream/data_source here since the connection will
be transferred to another shard, and these objects are deeply intertwined
with the shard, but we can use pollable_fd which is just a wrapper around
a file descriptor. We don't need to use the equivalent of read_exactly()
because the protocol guarantees the header will be sent in one packet
and not be fragmented.

In this patch, we only parse the protocol header and extract the remote
address. Changing connected_socket::local_address() and
connected_socket::remote_address() to reflect the proxy data is done later.