tls: run shutdown on main scheduling group
This prevents a nullptr-access UB during when the scheduling group
`tls::session::close()` was running on is destroyed before the
background fibre completes.

It is possible that during application shutdown, the scheduling group is
destroyed before the 10s timeout in the background fibre of tls session
shutdown completes. In this case, the timer would trigger UB while
trying to register the timeout callback onto the already destroyed
scheduling group's task queue.

To fix this, we move the tls shutdown background fibre to the main
scheduling fibre, which has a sufficiently long lifetime to complete the
task.

We do this for both the OpenSSL and GnuTLS stacks.