reactor: assert task queue exists on add_task
Assert that the task queue of the scheduling group exists at the time
the task is added to the reactor.

This prevents UB caused by nullptr access when a task is scheduled onto
an already-destroyed scheduling group. This can happen, for example,
when a timer-expiry-triggered task tries to schedule a task with a
scheduling group that has been destroyed in the meantime.